fix(scatter): fix issue breaking scale transitions
A contributor made a change in commit fca4c2d that seems to prevent the code
in the scatter model to reach a logical branch responsible for making transitions
between data sets with different ranges work properly for line charts.

The issue is only visible when transitioning from a data set with a small range
to one with a large range. For other cases the issue may pass unnoticed.

My assumption is that there's a mistake on the branch of the code that was not
reached before and that clamping the Y scale of the chart is the right thing to
do here in order to have transitions work they used to.